+++ srcs/perApp/messageSender/messageSenderFunctions.go
@@ -6,7 +6,7 @@ import (
 	tgbotapi "github.com/go-telegram-bot-api/telegram-bot-api"
 )
 
-func DiscordSendMessage(text string, dest int) error {
+func DiscordSendMessage(text string, dest uint64) error {
 	dmChan, err := definitions.Socket.DiscordSession.UserChannelCreate(fmt.Sprintf("%d", dest))
 	if err != nil {
 		return err
@@ -18,7 +18,7 @@ func DiscordSendMessage(text string, dest int) error {
 	return nil
 }
 
-func TelegramSendMessage(text string, dest int) error {
+func TelegramSendMessage(text string, dest uint64) error {
 	msg := tgbotapi.NewMessage(int64(dest), text)
 	_, err := definitions.Socket.TelegramSession.Send(msg)
 	if err != nil {
